+++ packages/web/src/app/[domain]/search/components/codePreviewPanel/index.tsx
@@ -32,9 +32,9 @@ export const CodePreviewPanel = ({
         queryKey: ["source", previewedFile, branch],
         queryFn: () => unwrapServiceError(
             getFileSource({
-                fileName: previewedFile.fileName.text,
-                repository: previewedFile.repository,
-                branch,
+                path: previewedFile.fileName.text,
+                repo: previewedFile.repository,
+                ref: branch,
             })
         ),
         select: (data) => {
@@ -42,7 +42,7 @@ export const CodePreviewPanel = ({
                 content: data.source,
                 filepath: previewedFile.fileName.text,
                 matches: previewedFile.chunks,
-                link: previewedFile.webUrl,
+                link: previewedFile.externalWebUrl,
                 language: previewedFile.language,
                 revision: branch ?? "HEAD",
             };